+++ packages/extension-polkagate/src/fullscreen/settings/partials/AiTransactionInfo.tsx
@@ -40,13 +40,13 @@ export default function AiTransactionInfo(): React.ReactElement {
   return (
     <Stack direction='column'>
       <Typography color='text.primary' fontSize='22px' m='45px 0 15px' sx={{ display: 'block', textAlign: 'left', textTransform: 'uppercase' }} variant='H-4'>
-        {t('AI Transaction Details')}
+        {t('AI Transaction Insights')}
       </Typography>
       <Stack alignItems='center' direction='row' gap='5px' justifyContent='center'>
         <MySwitch
           checked={enabled}
           columnGap='8px'
-          label={t('Enable AI transaction details on dapp signing')}
+          label={t('Enable AI transaction insights during dApp signing')}
           onChange={onChange}
         />
         {enabled &&