+++ packages/extension-polkagate/src/popup/signing/ledger/LedgerButtons.tsx
@@ -15,7 +15,7 @@ interface Props {
   error: string | null | undefined;
   disabled?: boolean;
   isBusy?: boolean | undefined;
-  ledgerWarning: string | null;
+  isMetadataProofReady?: boolean;
   ledgerLocked: boolean;
   onRefresh: () => void;
   onSignLedger: () => void;
@@ -23,20 +23,17 @@ interface Props {
   style?: React.CSSProperties;
 }
 
-function LedgerButtons({ disabled, error, isBusy, ledgerLocked, ledgerWarning, onCancel, onRefresh, onSignLedger, style = {} }: Props): React.ReactElement<Props> {
+function LedgerButtons({ disabled, error, isBusy, isMetadataProofReady, ledgerLocked, onCancel, onRefresh, onSignLedger, style = {} }: Props): React.ReactElement<Props> {
   const { t } = useTranslation();
   const isBlueish = useIsBlueish();
 
   return (
     <Grid container sx={{ bottom: 0, position: 'absolute', ...style }}>
-      {!!ledgerWarning &&
-        <LedgerErrorMessage error={ledgerWarning} />
-      }
       {error &&
         <LedgerErrorMessage error={error} />
       }
-      {!error && !ledgerWarning &&
-        <Grid alignItems='center' columnGap='5px' container item sx={{ mb: '25px' }}>
+      {!error &&
+        <Grid alignItems='center' columnGap='5px' container item sx={{ mb: '15px' }}>
           <ColorSwatch color={isBlueish ? '#596AFF' : '#674394'} size='24px' variant='Bold' />
           <Typography color={isBlueish ? '#809ACB' : '#AA83DC'} sx={{ textAlign: 'left', width: '90%' }} variant='B-4'>
             {t('This is a ledger account. To complete this transaction, use your ledger')}
@@ -45,13 +42,18 @@ function LedgerButtons({ disabled, error, isBusy, ledgerLocked, ledgerWarning, o
       }
       <DecisionButtons
         cancelButton
-        disabled={disabled}
+        disabled={ disabled || isMetadataProofReady === false }
         divider
         flexibleWidth
         isBusy={isBusy}
         onPrimaryClick={ledgerLocked || error ? onRefresh : onSignLedger}
         onSecondaryClick={onCancel}
-        primaryBtnText={ledgerLocked || error ? t('Refresh') : t('Sign on Ledger')}
+        primaryBtnText={
+          ledgerLocked || error
+          ? t('Refresh')
+          : !isMetadataProofReady
+          ? t('Loading proof...')
+          : t('Sign on Ledger')}
         secondaryBtnText={t('Cancel')}
       />
     </Grid>